# packages/contract/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Video Agent Public API",
    "description": "Cookie authentication and canonical-origin writes. Product JSON bodies are limited to 65536 bytes. Runtime constraints and official AG-UI custom values are not fully representable as JSON Schema.",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {
...
      "UUIDInput": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "string",
        "format": "uuid",
        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
      },
...
      "ThreadsResponse": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "threads": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "threadID": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "title": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 160,
                  "pattern": "^[^\\u0000\\ud800-\\udfff]*$",
                  "description": "Trimmed before validation: 1–160 UTF-16 code units after removing surrounding whitespace. Foreign validators may count Unicode code points instead."
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                },
                "archivedAt": {
                  "anyOf": [
                    {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                    },
                    {
                      "type": "null"
                    }
                  ]
                }
              },
              "required": [
                "threadID",
                "title",
                "createdAt",
                "archivedAt"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "threads"
        ],
        "additionalProperties": false
      },
...
      "EmptyRequestInput": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {},
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "Native Better Auth signed cookie; production cookie naming follows SDK configuration."
      }
    }
  },
  "paths": {
    "/api/logout": {
      "post": {
        "operationId": "logout",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyRequestInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session durably revoked and SDK cookies expired"
          },
          "403": {
            "description": "Untrusted origin"
          },
          "408": {
            "description": "Request body timed out; retry same request"
          },
          "413": {
            "description": "Request body exceeds byte limit"
          },
          "415": {
            "description": "Expected JSON"
          },
          "503": {
            "description": "Collection stopped or transport failed"
          }
        }
      }
    },
...
    "/api/threads": {
      "get": {
        "operationId": "listThreads",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Owned threads",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Untrusted origin"
          },
          "404": {
            "description": "Not found (including foreign identifiers)"
          },
          "408": {
            "description": "Request body timed out; retry same request"
          },
          "409": {
            "description": "Conflicting intent or archived thread"
          },
          "413": {
            "description": "Request body exceeds byte limit"
          },
          "415": {
            "description": "Expected JSON request"
          },
          "503": {
            "description": "Unavailable; retry after recovery"
          }
        }
      },
      "post": {
        "operationId": "createThread",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Exact replay",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadResponse"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Untrusted origin"
          },
          "404": {
            "description": "Not found (including foreign identifiers)"
          },
          "408": {
            "description": "Request body timed out; retry same request"
          },
          "409": {
            "description": "Conflicting intent or archived thread"
          },
          "413": {
            "description": "Request body exceeds byte limit"
          },
          "415": {
            "description": "Expected JSON request"
          },
          "503": {
            "description": "Unavailable; retry after recovery"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThreadCreationInput"
              }
            }
          }
        }
      }
    },
    "/api/threads/{threadID}": {
      "get": {
        "operationId": "getThread",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Owned thread",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Untrusted origin"
          },
          "404": {
            "description": "Not found (including foreign identifiers)"
          },
          "408": {
            "description": "Request body timed out; retry same request"
          },
          "409": {
            "description": "Conflicting intent or archived thread"
          },
          "413": {
            "description": "Request body exceeds byte limit"
          },
          "415": {
            "description": "Expected JSON request"
          },
          "503": {
            "description": "Unavailable; retry after recovery"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "threadID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UUIDInput"
            }
          }
        ]
      },
      "patch": {
        "operationId": "updateThread",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Untrusted origin"
          },
          "404": {
            "description": "Not found (including foreign identifiers)"
          },
          "408": {
            "description": "Request body timed out; retry same request"
          },
          "409": {
            "description": "Conflicting intent or archived thread"
          },
          "413": {
            "description": "Request body exceeds byte limit"
          },
          "415": {
            "description": "Expected JSON request"
          },
          "503": {
            "description": "Unavailable; retry after recovery"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "threadID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UUIDInput"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThreadUpdateInput"
              }
            }
          }
        }
      }
    },
    "/api/threads/{threadID}/archive": {
      "post": {
        "operationId": "archiveThread",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Archived",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Untrusted origin"
          },
          "404": {
            "description": "Not found (including foreign identifiers)"
          },
          "408": {
            "description": "Request body timed out; retry same request"
          },
          "409": {
            "description": "Conflicting intent or archived thread"
          },
          "413": {
            "description": "Request body exceeds byte limit"
          },
          "415": {
            "description": "Expected JSON request"
          },
          "503": {
            "description": "Unavailable; retry after recovery"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "threadID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UUIDInput"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyRequestInput"
              }
            }
          }
        }
      }
    },
    "/api/threads/{threadID}/messages": {
      "get": {
        "operationId": "listMessages",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Public snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessagesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Untrusted origin"
          },
          "404": {
            "description": "Not found (including foreign identifiers)"
          },
          "408": {
            "description": "Request body timed out; retry same request"
          },
          "409": {
            "description": "Conflicting intent or archived thread"
          },
          "413": {
            "description": "Request body exceeds byte limit"
          },
          "415": {
            "description": "Expected JSON request"
          },
          "503": {
            "description": "Unavailable; retry after recovery"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "threadID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UUIDInput"
            }
          }
        ]
      },
      "post": {
        "operationId": "submitMessage",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "202": {
            "description": "Durably accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageAccepted"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Untrusted origin"
          },
          "404": {
            "description": "Not found (including foreign identifiers)"
          },
          "408": {
            "description": "Request body timed out; retry same request"
          },
          "409": {
            "description": "Conflicting intent or archived thread"
          },
          "413": {
            "description": "Request body exceeds byte limit"
          },
          "415": {
            "description": "Expected JSON request"
          },
          "503": {
            "description": "Unavailable; retry after recovery"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "threadID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UUIDInput"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageSubmissionInput"
              }
            }
          }
        }
      }
    },
    "/api/threads/{threadID}/runs/{runID}/cancel": {
      "post": {
        "operationId": "cancelRun",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "202": {
            "description": "Durably accepted cancellation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancellationAccepted"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Untrusted origin"
          },
          "404": {
            "description": "Not found (including foreign identifiers)"
          },
          "408": {
            "description": "Request body timed out; retry same request"
          },
          "409": {
            "description": "Conflicting intent or archived thread"
          },
          "413": {
            "description": "Request body exceeds byte limit"
          },
          "415": {
            "description": "Expected JSON request"
          },
          "503": {
            "description": "Unavailable; retry after recovery"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "threadID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UUIDInput"
            }
          },
          {
            "in": "path",
            "name": "runID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UUIDInput"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCancellationInput"
              }
            }
          }
        }
      }
    },
...
  }
}